# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "koora API",
        "contact": {
            "name": "Ahmed Magdy and abouzaid",
            "url": "https://www.google.com"
        },
        "version": "1.0"
    },
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/Players/All-Players": {
            "get": {
                "tags": [
                    "Player"
                ],
                "description": "get all Players",
                "operationId": "GET_ALL_Teams",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Players/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "description": "get Players with it's id",
                "operationId": "GET_Player",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Remove player from database",
                "operationId": "deleteplayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "player ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/Players/Create": {
            "post": {
                "tags": [
                    "Player"
                ],
                "description": "post Create Player form ",
                "operationId": "POST_Create_Player",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player_creation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/Players/Update/{id}": {
            "put": {
                "tags": [
                    "Player"
                ],
                "description": "put request for update Player name",
                "operationId": "PUT_UPDATE_Player_name",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player_creation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Seasons/{season_id}/RegisteredTeam/{team_id}/RegisteredPlayer/{id}": {
            "get": {
                "tags": [
                    "Registered Players"
                ],
                "description": "get singe team player",
                "operationId": "findPlayer",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season_id",
                        "in": "path",
                        "description": "Season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "team_id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Registered Players"
                ],
                "description": "Delete player",
                "operationId": "deletePlayer",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "season_id",
                        "in": "path",
                        "description": "Season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "team_id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "SUCCESSFULLY Deleted"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/api/Teams/All-Teams": {
            "get": {
                "tags": [
                    "Team"
                ],
                "description": "get all Teams",
                "operationId": "GET_ALL_Teams",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESSFULLY DONE"
                    },
                    "401": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "player_creation": {
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Mohamed Salah"
                },
                "player_position": {
                    "type": "string",
                    "example": "GK"
                },
                "team_id": {
                    "type": "integer",
                    "example": "5"
                },
                "country_id": {
                    "type": "integer",
                    "example": "2"
                }
            }
        }
    }
}